feat: return a Color object instead of specific values for each slider
Make it so that we get a Color object from the sliders instead of a value representing the slider
position. This makes it possible to use just one universal handler no matter which slider we want to
use.

BREAKING CHANGE: Each slider will be returning a Color object, not a number.

re #9